rename file holding common test dependency
